Move rule out to its own file.
